ADD GIMP TOOL AND DUMMY DOCKER COMPOSE FILES TO THE COMPOSE FILE PATH
Adding the GIMP tool's Docker Compose file is obviously so that the tool
can be used by this project. The dummy Docker Compose file takes a
little more explanation.

The dummy.yml file sets the directory that relative paths in subsequent
Docker Compose files are relative to. It must come first in the Docker
Compose file path to do this.

The main project docker-compose.yml file overrides the value of the
WORDPRESS_CONFIG_EXTRA environment variable that is set in the Docker
Compose file from tools/wordpress and must come after that Docker
Compose file in the COMPOSE_FILE paths to do so, so it can't be first in
the COMPOSE_FILE paths, hence the need for the dummy.yml Docker Compose
file.